make link travel times be consistent with BeamPath and BeamLeg durations by using Double at link level, also start work on resolving conflicts in pass schedules

+++ src/main/scala/beam/agentsim/agents/vehicles/BeamVehicle.scala
@@ -321,7 +321,7 @@ object BeamVehicle {
         )
     } else {
       val linkIds = beamLeg.travelPath.linkIds.drop(1)
-      val linkTravelTimes: IndexedSeq[Int] = beamLeg.travelPath.linkTravelTime.drop(1)
+      val linkTravelTimes: IndexedSeq[Double] = beamLeg.travelPath.linkTravelTime.drop(1)
       // generate the link arrival times for each link ,by adding cumulative travel times of previous links
 //      val linkArrivalTimes = linkTravelTimes.scan(beamLeg.startTime)((enterTime,duration) => enterTime + duration).dropRight(1)
 //      val nextLinkIds = linkIds.takeRight(linkIds.size - 1)
