+++ src/main/scala/beam/agentsim/events/PathTraversalEvent.scala
@@ -27,7 +27,7 @@ case class PathTraversalEvent(
   mode: BeamMode,
   legLength: Double,
   linkIds: IndexedSeq[Int],
-  linkTravelTime: IndexedSeq[Int],
+  linkTravelTime: IndexedSeq[Double],
   startX: Double,
   startY: Double,
   endX: Double,
@@ -218,8 +218,8 @@ object PathTraversalEvent {
     val linkIdsAsStr = Option(attr(ATTRIBUTE_LINK_IDS)).getOrElse("")
     val linkIds: IndexedSeq[Int] = if (linkIdsAsStr == "") IndexedSeq.empty else linkIdsAsStr.split(",").map(_.toInt)
     val linkTravelTimeStr = attr.getOrElse(ATTRIBUTE_LINK_TRAVEL_TIME, "")
-    val linkTravelTime: IndexedSeq[Int] =
-      if (linkTravelTimeStr == "") IndexedSeq.empty else linkTravelTimeStr.split(",").map(_.toInt)
+    val linkTravelTime: IndexedSeq[Double] =
+      if (linkTravelTimeStr == "") IndexedSeq.empty else linkTravelTimeStr.split(",").map(_.toDouble)
     val startX: Double = attr(ATTRIBUTE_START_COORDINATE_X).toDouble
     val startY: Double = attr(ATTRIBUTE_START_COORDINATE_Y).toDouble
     val endX: Double = attr(ATTRIBUTE_END_COORDINATE_X).toDouble
